remove interactive mode from event creation
was a good idea but the script should run as simple as possible, read: on the server

Also, the logic is actually quite simple and doesn't need user knowledge: Add event if there exists no event yet. If already an event exists and the name is different, also add it. Otherwise, don't add it.